Use saasUseExtension syntax
Without the word "extension", the variable is freeform and need to have
special code to handle it, like for Scribunto. The goal is to be able
to load pack of extensions.

In another hand, when the setting needs to be parsed automatically,
"saasUseExtension" and "saasUseSkin" prefixes are required.

This is a follow-up for 91ebcb2ce45e.

Differential Revision: https://devcentral.nasqueron.org/D3855